// src/common/silex-web/editor/api/api.ts
export interface JSONSchema {
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    $id?: string;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    $schema?: string;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    $ref?: string;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    $comment?: string;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    title?: string;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    description?: string;
    /**
     *
     * @type {any}
     * @memberof JSONSchema
     */
    _default?: any;
    /**
     *
     * @type {boolean}
     * @memberof JSONSchema
     */
    readOnly?: boolean;
    /**
     *
     * @type {Array<any>}
     * @memberof JSONSchema
     */
    examples?: Array<any>;
    /**
     *
     * @type {number}
     * @memberof JSONSchema
     */
    multipleOf?: number;
    /**
     *
     * @type {number}
     * @memberof JSONSchema
     */
    maximum?: number;
    /**
     *
     * @type {number}
     * @memberof JSONSchema
     */
    exclusiveMaximum?: number;
    /**
     *
     * @type {number}
     * @memberof JSONSchema
     */
    minimum?: number;
    /**
     *
     * @type {number}
     * @memberof JSONSchema
     */
    exclusiveMinimum?: number;
    /**
     *
     * @type {number}
     * @memberof JSONSchema
     */
    maxLength?: number;
    /**
     *
     * @type {NonNegativeIntegerDefault0}
     * @memberof JSONSchema
     */
    minLength?: NonNegativeIntegerDefault0;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    pattern?: string;
    /**
     *
     * @type {JSONSchema}
     * @memberof JSONSchema
     */
    additionalItems?: JSONSchema;
    /**
     *
     * @type {any}
     * @memberof JSONSchema
     */
    items?: any;
    /**
     *
     * @type {number}
     * @memberof JSONSchema
     */
    maxItems?: number;
    /**
     *
     * @type {NonNegativeIntegerDefault0}
     * @memberof JSONSchema
     */
    minItems?: NonNegativeIntegerDefault0;
    /**
     *
     * @type {boolean}
     * @memberof JSONSchema
     */
    uniqueItems?: boolean;
    /**
     *
     * @type {JSONSchema}
     * @memberof JSONSchema
     */
    contains?: JSONSchema;
    /**
     *
     * @type {number}
     * @memberof JSONSchema
     */
    maxProperties?: number;
    /**
     *
     * @type {NonNegativeIntegerDefault0}
     * @memberof JSONSchema
     */
    minProperties?: NonNegativeIntegerDefault0;
    /**
     *
     * @type {StringArray}
     * @memberof JSONSchema
     */
    required?: StringArray;
    /**
     *
     * @type {JSONSchema}
     * @memberof JSONSchema
     */
    additionalProperties?: JSONSchema;
    /**
     *
     * @type {{ [key: string]: JSONSchema; }}
     * @memberof JSONSchema
     */
    definitions?: { [key: string]: JSONSchema };
    /**
     *
     * @type {{ [key: string]: JSONSchema; }}
     * @memberof JSONSchema
     */
    properties?: { [key: string]: JSONSchema };
    /**
     *
     * @type {{ [key: string]: JSONSchema; }}
     * @memberof JSONSchema
     */
    patternProperties?: { [key: string]: JSONSchema };
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof JSONSchema
     */
    dependencies?: { [key: string]: any };
    /**
     *
     * @type {JSONSchema}
     * @memberof JSONSchema
     */
    propertyNames?: JSONSchema;
    /**
     *
     * @type {any}
     * @memberof JSONSchema
     */
    _const?: any;
    /**
     *
     * @type {Array<string>}
     * @memberof JSONSchema
     */
    _enum?: Array<string>;
    /**
     *
     * @type {any}
     * @memberof JSONSchema
     */
    type?: any;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    format?: string;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    contentMediaType?: string;
    /**
     *
     * @type {string}
     * @memberof JSONSchema
     */
    contentEncoding?: string;
    /**
     *
     * @type {JSONSchema}
     * @memberof JSONSchema
     */
    _if?: JSONSchema;
    /**
     *
     * @type {JSONSchema}
     * @memberof JSONSchema
     */
    then?: JSONSchema;
    /**
     *
     * @type {JSONSchema}
     * @memberof JSONSchema
     */
    _else?: JSONSchema;
    /**
     *
     * @type {SchemaArray}
     * @memberof JSONSchema
     */
    allOf?: SchemaArray;
    /**
     *
     * @type {SchemaArray}
     * @memberof JSONSchema
     */
    anyOf?: SchemaArray;
    /**
     *
     * @type {SchemaArray}
     * @memberof JSONSchema
     */
    oneOf?: SchemaArray;
    /**
     *
     * @type {JSONSchema}
     * @memberof JSONSchema
     */
    not?: JSONSchema;
}

/**
 *
 * @export
 * @interface ModelError
 */
export interface ModelError {
    /**
     *
     * @type {number}
     * @memberof ModelError
     */
    code: number;
    /**
     *
     * @type {string}
     * @memberof ModelError
     */
    message: string;
}

/**
 *
 * @export
 * @interface NonNegativeIntegerDefault0
 */
export interface NonNegativeIntegerDefault0 {}

/**
 *
 * @export
 * @interface Option
 */
export interface Option extends BaseWidget {
    /**
     *
     * @type {UISchema}
     * @memberof Option
     */
    uiSchema: UISchema;
}

/**
 *
 * @export
 * @interface SchemaArray
 */
export interface SchemaArray extends Array<JSONSchema> {}

/**
 *
 * @export
 * @enum {string}
 */
export enum SimpleTypes {
    Array = "array",
    Boolean = "boolean",
    Integer = "integer",
    Null = "null",
    Number = "number",
    Object = "object",
    String = "string",
}

/**
 *
 * @export
 * @interface StringArray
 */
export interface StringArray extends Array<string> {}

/**
 *
 * @export
 * @interface Template
 */
export interface Template extends Entire {}

/**
 *
 * @export
 * @interface TemplateIntance
 */
export interface TemplateIntance {
    /**
     *
     * @type {string}
     * @memberof TemplateIntance
     */
    aa?: string;
}

/**
 *
 * @export
 * @interface UISchema
 */
export interface UISchema {
    /**
     *
     * @type {Uiwidget}
     * @memberof UISchema
     */
    ui$widget?: Uiwidget;
    /**
     *
     * @type {string}
     * @memberof UISchema
     */
    classNames?: string;
    /**
     *
     * @type {any}
     * @memberof UISchema
     */
    ui$options?: any;
    /**
     *
     * @type {boolean}
     * @memberof UISchema
     */
    ui$disabled?: boolean;
    /**
     *
     * @type {boolean}
     * @memberof UISchema
     */
    ui$readonly?: boolean;
    /**
     *
     * @type {Array<string>}
     * @memberof UISchema
     */
    ui$order?: Array<string>;
    /**
     *
     * @type {Array<string>}
     * @memberof UISchema
     */
    ui$enumDisabled?: Array<string>;
    /**
     *
     * @type {string}
     * @memberof UISchema
     */
    ui$rootFieldId?: string;
    /**
     *
     * @type {string}
     * @memberof UISchema
     */
    ui$help?: string;
    /**
     *
     * @type {string}
     * @memberof UISchema
     */
    ui$title?: string;
    /**
     *
     * @type {string}
     * @memberof UISchema
     */
    ui$description?: string;
    /**
     *
     * @type {boolean}
     * @memberof UISchema
     */
    ui$autofocus?: boolean;
    /**
     *
     * @type {string}
     * @memberof UISchema
     */
    ui$placeholder?: string;
    /**
     *
     * @type {any}
     * @memberof UISchema
     */
    ui$emptyValue?: any;
    /**
     *
     * @type {boolean}
     * @memberof UISchema
     */
    ui$print?: boolean;
}

/**
 *
 * @export
 * @interface Uioptions
 */
export interface Uioptions {}

/**
 *
 * @export
 * @enum {string}
 */
export enum Uiwidget {
    Checkbox = "checkbox",
    Radio = "radio",
    Select = "select",
    Hidden = "hidden",
    Text = "text",
    Password = "password",
    Email = "email",
    Hostname = "hostname",
    Ipv4 = "ipv4",
    Ipv6 = "ipv6",
    Uri = "uri",
    DataUrl = "data-url",
    Textarea = "textarea",
    Date = "date",
    Datetime = "datetime",
    DateTime = "date-time",
    AltDate = "alt-date",
    AltDatetime = "alt-datetime",
    Color = "color",
    Updown = "updown",
    Range = "range",
    File = "file",
    Checkboxes = "checkboxes",
    Files = "files",
}